+++ src/pages/about.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import About from "@screens/about";
+import About from "@src/src/screens/about";
 
 const AboutPage = () => {
   return <About />;
