+++ src/screens/about/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Link from "next/link";
-import Layout from "@components/layout";
+import Layout from "@src/src/components/layout";
 
 const About = () => {
   return (
