+++ src/screens/home/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Link from "next/link";
 import Head from "next/head";
-import Layout from "@components/layout";
+import Layout from "@src/src/components/layout";
 import { useTranslation } from "@src/i18n";
 
 const Home = () => {
