+++ src/static/hero.scss
