+++ tsconfig.json
@@ -23,10 +23,10 @@
         "./*"
       ],
       "@components/*": [
-        "components/*"
+        "src/components/*"
       ],
       "@screens/*": [
-        "screens/*"
+        "src/screens/*"
       ],
       "@public/*": [
         "public/*"
